Refactor how eval is called and used
Renamed the function to _eval_in_program_context() to better reflect what it
is doing.  It's now a normal function call instead of expecting args in
the $DB::eval_string variable.

Debugger clients request for an eval with the $db->eval() method.  All the
evals are done the next time through the client loop.  Results are delivered
by a callback.