Fix scrolling issue with the main code pane
It was initially drawing the main code pane too tall causing it to extend
past the bottom of the browser window, so the main window's scrollbar would
activate.

An element (in this case, the div containing the code) has no offset().top
until you attach it to the documen, and so was calculating the pinned height
incorrectly.