Perl var popovers are more robust
If you move the pointer in and out of several perl vars one after another,
it could draw the popover for the last variable, but the contents are actually
for one of the prior variables.

This fixes the problem by moving the hovered-perl-var class from the text
element representing the perl var on mouseout.  When we get the results back
from getting the var's value, check to see if the element we're retrieving
the value for still has the hovered-perl-var class (meaning the pointer is
still over it)