Child process listen socket tries to use original port + 1
After a fork() the child process tries to open it's own listen socket by
bumping the parent process's port up by 1.  It uses a window of 20 to keep
trying to find an open one.

If it can't, it gives up and lets the system pick one for us.

This fixes #101