Hiliting double-quoted strings - partial fix
Tries to wrap a span.next-statement around something containing a double-
quoted string, as in
    print "got $count things\n";

It's encoded by a pp_stringify, and deparses as
    print qq(got $count things\n);
But the actual quotes from the source can't be determined, so we have to make
a regex that'll match any of the quote types
    qq( qq[ qq{ qq/ "

The matching part of the replace() is working, but the replacement isn't quite
correct.  The UI is only hiliting up to the first quote
    print "

This will need some more work.