Add argument 'launch'
If specified, this argument attempts to launch a browser on the
debugger's port. The value of the argument is the name of the browser to
launch. If the argument is omitted, Browser::Open is loaded, and its
favored browser is used. If the attempt to load Browser::Open fails, an
exception is raised.